attendance_backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Machine Learning & Computer Vision
opencv-python==4.8.0.76
torch==2.1.0
torchvision==0.16.0
numpy==1.24.3
Pillow==10.0.0

# Face Detection & Recognition
ultralytics==8.0.0  # YOLOv8 for face detection
facenet-pytorch==2.5.0  # FaceNet implementation
deepface==0.0.75  # Alternative to facenet-pytorch (includes RetinaFace)
retinaface==0.0.13  # Face detector

# Vector Search & Indexing
faiss-cpu==1.7.4  # Facebook's similarity search library
scikit-learn==1.3.1  # For KD-tree and preprocessing

# Database & Persistence
firebase-admin==6.2.0
python-dotenv==1.0.0

# Data Processing
pandas==2.0.3
scipy==1.11.3

# Async & Concurrency
aiofiles==23.2.1
asyncio-timeout==4.0.3

# Logging & Monitoring
python-json-logger==2.0.7

# Utilities
pydantic==2.4.2
pydantic-settings==2.0.3
requests==2.31.0

# Development (Optional)
pytest==7.4.2
pytest-asyncio==0.21.1
black==23.10.1
flake8==6.1.0
mypy==1.6.1
